:hammer: add authorization, jwt settings to .env, post --> user relations

+++ controllers/postController.js
@@ -1,5 +1,6 @@
+const mongoose = require("mongoose");
 const Post = require("../models/Post");
-const path = require("path");
+const User = require("../models/User");
 const { validationResult } = require("express-validator");
 
 exports.getAll = (req, res, next) => {
@@ -37,25 +38,35 @@ exports.create = (req, res, next) => {
       }),
     });
   }
-
-  // find creator by reading the jwt
-  const creator = "";
-
-  const post = new Post({
-    title: req.body.title,
-    content: req.body.content,
-    imageUrl: req.file.path,
-    creator: creator,
-    createdAt: Date.now(),
-  });
-
-  post.save().then((post) => {
-    res.json({
-      success: true,
-      message: "Post Created !",
-      data: post,
+  // find creator by reading the req._id
+  User.findById(req.userId)
+    .then((user) => {
+      if (!user) {
+        return res.json({
+          success: false,
+          message: "user doest not exist!",
+        });
+      }
+      // create a new post object with creator data
+      const post = new Post({
+        title: req.body.title,
+        content: req.body.content,
+        imageUrl: req.file.path,
+        creator: user._id,
+        createdAt: Date.now(),
+      });
+      // save the post
+      post.save().then((post) => {
+        res.json({
+          success: true,
+          message: "post created !",
+          data: post,
+        });
+      });
+    })
+    .catch((err) => {
+      console.log(err);
     });
-  });
 };
 
 exports.update = (req, res, next) => {
@@ -66,6 +77,13 @@ exports.update = (req, res, next) => {
           success: false,
           message: "post does not exist!",
         });
+      }
+
+      if(post.creator.toString() !== req.userId.toString()) {
+        return res.json({
+          success: false,
+          message: 'unauthorized to edit this post'
+        })
       }
 
       post.title = req.body.title;
@@ -91,6 +109,14 @@ exports.delete = (req, res, next) => {
           message: "post does not exist!",
         });
       }
+
+      if (post.creator.toString() !== req.userId.toString()) {
+        return res.json({
+          success: false,
+          message: "unauthorized to delete this post",
+        });
+      }
+
       return post.delete();
     })
     .then((post) => {
